+++ packages/psychart/example/Pump Curve/src/index.js
@@ -9,15 +9,23 @@ const nData = 100;
 // Create a custom pump chart
 const pumpchart = new Pumpchart({
     curve: { // Define performance curves as a function of `q` (flow rate)
-        pump: '30-(q/100)^2',
-        system: '5+(q/75)^2',
+        pump: {
+            maxFlow: 3000,
+            maxHead: 35,
+        },
+        system: {
+            static: 5,
+            friction: 2e-6,
+        },
     },
     speed: {
         max: 60, // Max pump speed (rpm)
-        operation: 55, // Operational speed (rpm)
+        operation: 45, // Operational speed (rpm)
         steps: [15, 30, 45], // Minor pump curves (rpm)
     },
+    density: 1,
     units: {
+        density: 'g/cm3',
         flow: 'gpm',
         head: 'psi',
         power: 'kW',
@@ -31,10 +39,14 @@ const pumpchart = new Pumpchart({
 
 // Plot `nData` number of data points
 for (let t = now; t < in1hr; t += onehr / nData) {
-    const flow = Math.random() * 300;
-    const head = Math.random() * 20;
-    pumpchart.plot({ flow: flow, head: head }, { timestamp: t });
+    const flow = Math.random() * 2500;
+    const head = Math.random() * 25;
+    const power = flow * head / 2000;
+    pumpchart.plot({ flow: flow, head: head, power: power }, { timestamp: t });
 }
+
+// pumpchart.clearData();
+// pumpchart.plot({ flow: 500, head: 150, power: 50 });
 
 // Append pumpchart to the document <body>
 document.body.append(pumpchart.getElement());